requirements.txt
aiohttp==3.8.4
aiohttp-apispec==2.2.3
marshmallow==3.19.0
pytest==7.3.1
pytest-aiohttp==1.0.4
PyYAML==6.0
SQLAlchemy==2.0.15
asyncpg==0.27.0
marshmallow-dataclass==8.5.14
alembic==1.11.1
aiohttp-cors==0.7.0
aiofiles==23.1.0